requirements.txt
wikipedia
streamlit
langchain
duckduckgo-search
openai
python-dotenv
langchain_experimental